changing table_size to size in alloc

+++ 0x1A-hash_tables/0-hash_table_create.c
@@ -1,13 +1,13 @@
 #include "hash_tables.h"
 /**
- * hash_table_create - create a hash table
+ * hash_table_create - creates a hash table
  * @size: size of hash table
  * Return: hash table
  */
 hash_table_t *hash_table_create(unsigned long int size)
 {
 	hash_table_t *table;
-	size_t i;
+	unsigned long int i;
 
 	table = (hash_table_t *) malloc(sizeof(hash_table_t));
 	if (table == NULL)
